refactor: split runtime control-plane and MCP tooling modules into packages (#49)
Convert four >500-line modules into same-named packages behind API-stable
re-export facades (no behavior change):

- raes_runtime/control_plane_api  -> _responses/_auth/_operation_routes/
  _workflow_routes/_participant_routes; version seam + app factory stay in
  the facade so test_version_classification's distribution_version patch holds.
- raes_runtime/workflow_result_contract_checks -> _models/_context/
  _step_checks/_history_checks; diagnostic order preserved.
- raes_mcp/tools/authoring -> tools/_templates/_helpers.
- raes_mcp/tools/inspection -> tools/_common/_summary/_elements/_references.

Drain the four entries from oversized_allowlist.yaml; retarget the ruff
per-file-ignores, requirement_order.yaml ownership root, and the
shared-semantic-integrity coverage path to the new package locations.
