+++ implementations/python/pyproject.toml
@@ -177,7 +177,9 @@ ignore = [
 "tests/**" = ["S101", "S105", "S106", "S108", "E402", "SIM105", "F841"]
 "packages/raes_contracts/contracts/__init__.py" = ["F403", "F405"]  # intentional package re-exports
 "packages/raes_cli/**" = ["B008"]      # typer requires function calls in defaults
-"packages/raes_runtime/control_plane_api.py" = ["B008"]  # fastapi Depends() pattern
 "packages/raes_runtime/control_plane.py" = ["S112"]      # intentional exception suppression
 "packages/raes/module_registry/__init__.py" = ["S310", "S202", "F401"]  # S310/S202: explicit OCI URL fetch + tar extract; F401: intentional facade re-exports
+"packages/raes_runtime/control_plane_api/__init__.py" = ["F401"]  # intentional facade re-exports (_receipt_response)
+"packages/raes_mcp/tools/authoring/__init__.py" = ["F401"]  # intentional facade re-export (register)
+"packages/raes_mcp/tools/inspection/__init__.py" = ["F401"]  # intentional facade re-export (register)
 "packages/raes_contracts/realization_envelope.py" = ["E402"]  # late import breaks a carrier dependency cycle
