+++ tools/policy/oversized_allowlist.yaml
@@ -11,11 +11,7 @@ files:
   - implementations/python/packages/raes_backend_libvirt/realization.py
   - implementations/python/packages/raes_backend_libvirt/techvault_native.py
   - implementations/python/packages/raes_contracts/workflow.py
-  - implementations/python/packages/raes_mcp/tools/authoring.py
-  - implementations/python/packages/raes_mcp/tools/inspection.py
   - implementations/python/packages/raes_operations/_evidence_run_artifact.py
-  - implementations/python/packages/raes_runtime/control_plane_api.py
-  - implementations/python/packages/raes_runtime/workflow_result_contract_checks.py
   - implementations/python/packages/raes/composition.py
   - implementations/python/packages/raes/orchestration.py
   - implementations/python/packages/raes/participant_behavior.py
